feature: MET-1118 finding 4 update modal compose report UI
