+++ src/components/StakingLifeCycle/DelegatorLifecycle/ReportComposerModal/ReportComposerModal.test.tsx
@@ -1,9 +1,7 @@
 import { screen } from "@testing-library/react";
-
 import { render } from "src/test-utils";
 
 import ReportComposerModal from ".";
-
 test("should render ReportComposerModal", async () => {
   render(<ReportComposerModal open={true} handleCloseModal={jest.fn()} />);
   expect(screen.getByTestId("report-compose-modal")).toBeInTheDocument();
